fix: Extracting DateTime from json in Mode Item
Previous apprach extracted the string from json for the key. Now, when
we extra DateTime as string, it's ToString() would be called. Then
again, the extracted DateTime was parsed to get DateTime out of it.

The problem with previous approach was that, ToString() would convert
date to local dateFormat. When DateTime.Parse is called, we would not
submit dateFormat of string date. In case of locales where dateFormat is
dd/mm/yy this would created problem.

New approach directly typecasts the extracted value from json to
DateTime. This approach not only resolves the dateFormat problem, it
also does so in less steps.

Reported-by: Sudarshan Devardekar <sudarshansmd@gmail.com>
Signed-off-by: Sudarshan Devardekar <sudarshansmd@gmail.com>