fix: Extracting DateTime from json in Model Item
Previous approach to get DateTime from json used to extract the string from json using its key, and then prase the string using DateTime.Parse(). When DateTime is extracted from json as string, the DateTime is converted to string by calling ToString() on DateTime. When ToString() is called on DateTime, it converts DateTime to local string date format.

The problem with previous apprach was that, DateTime.ToString() would
convert the date to locale dateFormat. When DateTime.Parse() is again
called on string date, dateFormat of string is not provided. This
creates problem in parsing dates with certain dateFormat viz. dd/mm/yy.

New approach directly typecasts the extracted value from json to
DateTime. This approach not only resolves the dateFormat problem, but it
also does so in less steps.

Reported-by: Sudarshan Devardekar <sudarshansmd@gmail.com>
Signed-off-by: Sudarshan Devardekar <sudarshansmd@gmail.com>